requirements.txt
pygame>=2.1.0
Pillow>=9.0.0
numpy>=1.21.0
